# api/admin/admin.openapi.yaml
openapi: 3.0.3
info:
  title: MEASIX Admin API
  version: 1.0.0
paths:
...
  /api/admin/v1/session:
    get:
      operationId: getSession
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSession'
        '401':
          $ref: '#/components/responses/Problem'
    delete:
      operationId: logoutAdmin
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
      parameters:
      - &id001
        name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
  /api/admin/v1/users:
    get:
      operationId: listUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
      parameters: &id006
      - &id004
        name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - &id005
        name: cursor
        in: query
        schema:
          type: string
    post:
      operationId: createUser
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      parameters:
      - *id001
  /api/admin/v1/users/{userId}:
    get:
      operationId: getUser
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
      parameters:
      - &id002
        name: userId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UserId'
    put:
      operationId: updateUser
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      parameters:
      - *id002
      - *id001
...
  /api/admin/v1/users/{userId}:set-password:
    post:
      operationId: setPassword
      responses:
        '204':
          description: Updated
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
      parameters:
      - *id002
      - *id001
  /api/admin/v1/users/{userId}/enrollments:
    post:
      operationId: createEnrollment
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEnrollmentResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnrollmentRequest'
      parameters:
      - *id002
      - *id001
  /api/admin/v1/users/{userId}/devices:
    get:
      operationId: listDevices
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePage'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
      parameters:
      - *id002
      - *id004
      - *id005
...
components:
  schemas:
...
    UserId:
      type: string
      pattern: ^usr_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
    DeviceId:
      type: string
      pattern: ^dev_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
    EnrollmentId:
      type: string
      pattern: ^enr_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
...
    InstallationId:
      type: string
      pattern: ^ins_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
...
    ActivationId:
      type: string
      pattern: ^act_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
    RequestId:
      type: string
      pattern: ^req_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
...
    Problem:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
        detail:
          type: string
        requestId:
          $ref: '#/components/schemas/RequestId'
        activationId:
          $ref: '#/components/schemas/ActivationId'
        targetManagedGeneration:
          type: integer
          minimum: 0
        currentDraftRevision:
          type: integer
          minimum: 0
        forwarded:
          type: boolean
...
    AdminUserSummary:
      type: object
      additionalProperties: false
      required:
      - userId
      - displayName
      - role
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        displayName:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - MEMBER
    AdminSession:
      type: object
      additionalProperties: false
      required:
      - user
      - csrfToken
      - expiresAt
      properties:
        user:
          $ref: '#/components/schemas/AdminUserSummary'
        csrfToken:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    User:
      type: object
      additionalProperties: false
      required:
      - userId
      - username
      - displayName
      - role
      - status
      - createdAt
      - updatedAt
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - MEMBER
        status:
          type: string
          enum:
          - ACTIVE
          - DISABLED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      additionalProperties: false
      required:
      - username
      - displayName
      - role
      properties:
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - MEMBER
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required:
      - username
      - displayName
      - role
      properties:
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - MEMBER
    SetPasswordRequest:
      type: object
      additionalProperties: false
      required:
      - newPassword
      properties:
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
    CreateEnrollmentRequest:
      type: object
      additionalProperties: false
      properties:
        expiresInSeconds:
          type: integer
          minimum: 60
          maximum: 3600
          default: 600
    CreateEnrollmentResponse:
      type: object
      additionalProperties: false
      required:
      - enrollmentId
      - code
      - expiresAt
      - platformUrl
      properties:
        platformUrl:
          type: string
          format: uri
          description: Configured public HTTP or HTTPS platform origin; never inferred from the request.
        enrollmentId:
          $ref: '#/components/schemas/EnrollmentId'
        code:
          type: string
        expiresAt:
          type: string
          format: date-time
    Device:
      type: object
      additionalProperties: false
      required:
      - deviceId
      - deviceName
      - userId
      - status
      - applicationState
      - targetManagedGeneration
      properties:
        deviceName:
          type: string
          description: Device name supplied during enrollment; display metadata, not authorization identity.
        applicationState:
          type: string
          enum: [UNKNOWN, PENDING, APPLIED, UNPUBLISHED]
          description: Last report of the current valid session, not an online or runtime admission assertion.
        targetManagedGeneration:
          type: integer
          minimum: 0
        appliedManagedGeneration:
          type: integer
          minimum: 1
        appliedReportedAt:
          type: string
          format: date-time
        deviceId:
          $ref: '#/components/schemas/DeviceId'
        userId:
          $ref: '#/components/schemas/UserId'
        installationId:
          $ref: '#/components/schemas/InstallationId'
        status:
          type: string
          enum:
          - ACTIVE
          - REVOKED
        appVersion:
          type: string
        lastSeenAt:
          type: string
          format: date-time
    UserPage:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        nextCursor:
          type: string
    DevicePage:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Device'
        nextCursor:
          type: string
...
  responses:
    Problem:
      description: Problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'